Added GIF recording code. Reimplemented pool-based updating. Increased simulation area.

+++ core/src/com/bigbass/sandpiles/sim/Cell.java
@@ -2,7 +2,7 @@ package com.bigbass.sandpiles.sim;
 
 public class Cell {
 	
-	public int val;
+	public float val;
 	
 	public Cell(){
 		val = 0;
